backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pymongo
